+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.4.1 / _Not released yet_
+# 1.5.0 / 2013-05-20
 
 - On squeeze use default mirror with /debian-backports suffix for backports
   repository ([GH-7])